CIV-14025 Added unit tests
